presentation: Use "with permissions" wrapper
Implement and use `WithCameraPermission` wrapper composable instead of
the convoluted impl. right now. This also removes the awkward need for
`MainScreen` and `MainScreenContent` now the prior is the latter.
