+++ app/src/main/java/xyz/lilsus/papp/presentation/main/components/RequestCameraPermission.kt
@@ -1,0 +1,41 @@
+package xyz.lilsus.papp.presentation.main.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.PermissionState
+import com.google.accompanist.permissions.shouldShowRationale
+
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun RequestCameraPermission(cameraPermission: PermissionState) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .wrapContentSize()
+            .widthIn(max = 480.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        val text = if (cameraPermission.status.shouldShowRationale) {
+            "We need your camera to scan QR codes. Please grant permission!"
+        } else {
+            "Hi there! Grant camera permission to get started."
+        }
+        Text(text, textAlign = TextAlign.Center)
+        Spacer(Modifier.height(16.dp))
+        Button(onClick = { cameraPermission.launchPermissionRequest() }) {
+            Text("Grant Camera Access")
+        }
+    }
+}
